More spline fit superclass code + wrong B-CINPACT.
Refactored parts of the spline fit superclass to accommodate B-CINPACT.
But my implementation of B-Spline approximation using CINPACT was based
on a misinterpetation of the paper. I thought the sinc was still used.
But only the exponential part is used when approximating B-Splines.
Mine could work in theory, but then I wouldn't be able to use the k and
support radius values the paper gave for approximating B-Splines.
I'd have to find my own, which is not a good use of time.