Removing erroneous chart and the cause of it.
Doing a "step" on JAV-aligned displacements does not make sense because
the displacements would require summation (and reframing to do so).
So in the LSTM branch, this led me to make an LSTM that looked like it
was getting super low errors, but it was not.
So I'm modifying a cherrypicked commit from that branch to remove the
erroneous chart and the "dangerous" param that led to it.